refactor: move thread runtime run lifecycle helpers
